+++ esphome/components/dali_light/dali_light.h
@@ -9,11 +9,11 @@ class DALILight : public Component {
   void loop() override;
 
   void set_dali_bus(esphome::dali_bus::DALIBusComponent *bus);
-  void set_light_id(int id);
+  void set_light_id(uint8_t id);
 
  protected:
   esphome::dali_bus::DALIBusComponent *bus_;
-  int light_id_;
+  uint8_t light_id_;
 };
 
 }  // namespace dali_light
